add TABLES accounts, entries & transfers

+++ db/migrations/000001_user_setup.up.sql
@@ -1,7 +1,3 @@
 CREATE TABLE "users" (
-		"id" BIGSERIAL PRIMARY KEY,
-		"email" varchar(256) UNIQUE NOT NULL,
-		"hashed_password" varchar(256) NOT NULL,
-		"created_at" timestamptz NOT NULL DEFAULT(now()),
-		"updated_at" timestamptz NOT NULL DEFAULT(now())
+    "id" SERIAL PRIMARY KEY, "email" varchar(256) UNIQUE NOT NULL, "hashed_password" varchar(256) NOT NULL, "created_at" timestamptz NOT NULL DEFAULT(now()), "updated_at" timestamptz NOT NULL DEFAULT(now())
 );
